StreamlitApp/requirements.txt
google-cloud-firestore~=2.5.3
google-cloud-core~=2.3.1
streamlit>=1.10.0
pandas~=1.4.3
numpy~=1.23.0
google-cloud-bigquery~=3.2.0
google-cloud-bigquery-storage~=2.13.2
pydeck~=0.7.1
google-auth~=2.9.0
toml~=0.10.2
